RAG/requirements.txt
transformers
ctransformers
langchain
langchain-community
torch
pypdf
sentence-transformers
gpt4all
faiss-gpu
openai
langchain-cohere
bs4
spacy
tf-keras
datasets
python-dotenv